+++ erde/cli/compile.lua
@@ -93,6 +93,9 @@ local function watch_files(cli)
 end
 
 return function(cli)
+  -- Do not include rewrite error code for production compiled code
+  C.REWRITE_ERRORS = false
+
   if #cli == 0 then
     table.insert(cli, '.')
   end