+++ erde/cli/run.lua
@@ -18,10 +18,7 @@ return function(cli, script_args)
   local ok, result = pcall(function()
     local source = read_file(cli.script)
 
-    local result = lib.__erde_internal_load_source__(source, cli.script, {
-      bitlib = cli.bitlib,
-      rewrite_errors = true,
-    })
+    local result = lib.__erde_internal_load_source__(source, cli.script)
 
     return result
   end)